requirements.txt
Jinja2>=3.1
jsonschema>=4.0
anthropic>=0.39
python-dotenv>=1.0
